app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.6.3",
    "pydantic>=2.11.7",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "tenacity>=9.1.2",
]



[tool.pyright]
venvPath = "."
venv = ".venv"
